iss #277 move plot_meas_station_on_gmap to gis.py
