+++ brightwind/__init__.py
@@ -1,5 +1,5 @@
 from .load.load import *
-from .load.station import MeasurementStation, plot_meas_loc_on_gmap
+from .load.station import MeasurementStation
 from .analyse import correlation as Correl
 from .analyse.shear import *
 from .analyse.analyse import *
@@ -7,6 +7,7 @@ from .analyse.plot import *
 from .transform.transform import *
 from .export.export import *
 from . import demo_datasets
+from .utils.gis import *
 # from .utils.utils import *
 
 __all__ = ['analyse', 'transform', 'export', 'load', 'demo_datasets']
