+++ brightwind/load/station.py
@@ -16,225 +16,11 @@
 
 
 from brightwind.load.load import _is_file
-from brightwind.utils import utils
 import numpy as np
 import pandas as pd
 import requests
 import json
 import copy
-import gmaps
-
-
-__all__ = ['MeasurementStation',
-           'plot_meas_loc_on_gmap']
-
-
-# def _flatten_sensor_dict(sensor):
-#     """
-#         Flatten the sensor dictionary retrieved from jason
-#         assigning all the sub-dictionaries to the main dictionary.
-#
-#         :param sensor: The sensor dictionary retrieved for a single configuration
-#                            option and meas_point id.
-#         :type sensor: dict
-#         :return: output
-#         :rtype: dict
-#
-#     """
-#     output = {key: value for key, value in sensor.items() if (type(value) != list) or (value == {})}
-#     for key, value in zip(sensor.keys(), sensor.values()):
-#         if type(value) == list:
-#             if key == 'calibration':
-#                 value = {key + "_" + k: v for k, v in value[0].items()}
-#             output.update(value)
-#     return output
-#
-#
-# def _rename_variables(input_dict, root_name):
-#     for var_to_rename in ['height_m', 'serial_number', 'update_at', 'notes']:
-#         if var_to_rename in list(input_dict.keys()):
-#             input_dict[root_name + '_' + var_to_rename] = input_dict.pop(var_to_rename)
-#     return input_dict
-#
-#
-# def _get_meas_points(meas_points):
-#     meas_points_flatten = []
-#     for meas_point in meas_points:
-#         #         meas_point = _flatten_meas_point_dict(meas_point)
-#         sen_configs = sorted(meas_point['sensor_config'], key=lambda i: i['date_from'])
-#         sen_configs = [_replace_none_date(_rename_variables(sen_config, 'sen_config')) for sen_config in sen_configs]
-#         sensors = [_replace_none_date(_rename_variables(_flatten_sensor_dict(sensor), 'sensor')) for sensor in
-#                    meas_point['sensor']]
-#         if meas_point['mounting_arrangement'] is not None:
-#             mounting_arrangements = [_replace_none_date(_rename_variables(mntg_arrang, 'mounting_arrangement'))
-#                                      for mntg_arrang in meas_point['mounting_arrangement']]
-#         else:
-#             mounting_arrangements = {}
-#
-#         date_from = [sen_config['date_from'] for sen_config in sen_configs]
-#         date_to = [sen_config['date_to'] for sen_config in sen_configs]
-#         for sensor in sensors:
-#             date_from.append(sensor['date_from'])
-#             date_to.append(sensor['date_to'])
-#         for mntg_arrang in mounting_arrangements:
-#             date_from.append(mntg_arrang['date_from'])
-#             date_to.append(mntg_arrang['date_to'])
-#
-#         date_from.extend(date_to)
-#         dates = np.unique(date_from)
-#         for i in range(len(dates) - 1):
-#             good_sen_config = {}
-#             for sen_config in sen_configs:
-#                 if (sen_config['date_from'] <= dates[i]) & (sen_config['date_to'] > dates[i]):
-#                     good_sen_config = sen_config.copy()
-#             if good_sen_config != {}:
-#                 for sensor in sensors:
-#                     if (sensor['date_from'] <= dates[i]) & (sensor['date_to'] > dates[i]):
-#                         good_sen_config.update(sensor)
-#                 for mntg_arrang in mounting_arrangements:
-#                     if (mntg_arrang['date_from'] <= dates[i]) & (mntg_arrang['date_to'] > dates[i]):
-#                         good_sen_config.update(mntg_arrang)
-#                 good_sen_config['date_to'] = dates[i + 1]
-#                 good_sen_config['date_from'] = dates[i]
-#                 good_sen_config.update(meas_point)
-#                 del good_sen_config['sensor_config']
-#                 del good_sen_config['sensor']
-#                 meas_points_flatten.append(good_sen_config)
-#     return meas_points_flatten
-#
-#
-# def _format_sensor_table(meas_points, table_type='full'):
-#     if table_type == 'full':
-#         header = ['name', 'measurement_units', 'oem', 'model',
-#                   'height_m', 'boom_orientation_deg', 'vane_dead_band_orientation_deg',
-#                   'date_from', 'date_to', 'connection_channel', 'sen_config_height_m', 'slope', 'offset',
-#                   'calibration_slope',
-#                   'calibration_offset']
-#         header_for_report = ['Instrument Name', 'Units', 'Sensor OEM', 'Sensor Model',
-#                              'Height [m]', 'Boom Orient. [deg, mag N]', 'Dead Band Orient. [deg, mag N]',
-#                              'Date From', 'Date To', 'Logger Channel', 'Logger Stated Height [m]', 'Logger Slope',
-#                              'Logger Offset', 'Calibration Slope',
-#                              'Calibration Offset']
-#     elif table_type == 'meas_points':
-#         header = ['name', 'measurement_type_id', 'height_m', 'boom_orientation_deg']
-#         header_for_report = ['Instrument Name', 'Measurement Type', 'Height [m]', 'Boom Orient. [deg, mag N]']
-#     elif table_type == 'speed_info':
-#         header = ['name', 'measurement_units', 'oem', 'model', 'sensor_serial_number',
-#                   'height_m', 'boom_orientation_deg',
-#                   'date_from', 'date_to', 'connection_channel', 'slope', 'offset',
-#                   'calibration_slope', 'calibration_offset', 'measurement_type_id']
-#         header_for_report = ['Instrument Name', 'Units', 'Sensor Make', 'Sensor Model', 'Serial No',
-#                              'Height [m]', 'Boom Orient. [deg, mag N]',
-#                              'Date From', 'Date To', 'Logger Channel', 'Logger Slope', 'Logger Offset',
-#                              'Calibration Slope', 'Calibration Offset', 'measurement_type_id']
-#     elif table_type == 'direction_info':
-#         header = ['name', 'measurement_units', 'oem', 'model', 'sensor_serial_number',
-#                   'height_m', 'boom_orientation_deg', 'vane_dead_band_orientation_deg',
-#                   'date_from', 'date_to', 'connection_channel', 'offset', 'measurement_type_id']
-#         header_for_report = ['Instrument Name', 'Units', 'Sensor Make', 'Sensor Model', 'Serial No',
-#                              'Height [m]', 'Boom Orient. [deg, mag N]', 'Dead Band Orient. [deg, mag N]',
-#                              'Date From', 'Date To', 'Logger Channel', 'Logger Offset', 'measurement_type_id']
-#
-#     sensors_table_report = pd.DataFrame(meas_points)
-#
-#     if any(elem not in sensors_table_report.columns for elem in header):
-#         ind_to_remove = [ind for ind, elem in enumerate(header) if elem not in sensors_table_report.columns]
-#         del header[ind_to_remove[0]]
-#         del header_for_report[ind_to_remove[0]]
-#
-#     sensors_table_report = pd.DataFrame(sensors_table_report[header])
-#     if table_type == 'speed_info':
-#         sensors_table_report = sensors_table_report[sensors_table_report['measurement_type_id'] == 'wind_speed']
-#         del sensors_table_report['measurement_type_id']
-#     if table_type == 'direction_info':
-#         sensors_table_report = sensors_table_report[sensors_table_report['measurement_type_id'] == 'wind_direction']
-#         del sensors_table_report['measurement_type_id']
-#
-#     if 'date_from' in sensors_table_report.columns:
-#         sensors_table_report['date_from'] = pd.to_datetime(sensors_table_report['date_from']).dt.strftime("%d-%b-%Y")
-#         sensors_table_report['date_to'] = pd.to_datetime(sensors_table_report['date_to']).dt.strftime("%d-%b-%Y")
-#
-#     sensors_table_report = sensors_table_report.replace({np.nan: '-', 'NaT': '-', '31-Dec-2100': '-'})
-#     sensors_table_report.rename(columns={k: h for k, h in zip(header, header_for_report)}, inplace=True)
-#     index_name = 'Instrument Name'
-#     sensors_table_report = sensors_table_report.set_index(index_name)
-#
-#     return sensors_table_report
-
-
-def _create_coord_dict(name, latitude, longitude):
-    return {'name': name, 'coords': (latitude, longitude)}
-
-
-def plot_meas_loc_on_gmap(meas_station, map_type='TERRAIN',
-                          zoom_level=9):
-    """
-    Visualise on Google Maps the location of one or more measurement locations.
-
-    Unfortunately, to use this function you must have a Google Maps API key which should be free. To get one, follow
-    the 'Get Started' instructions on the Google Maps Platform or go to the 'Credentials' section:
-
-    https://cloud.google.com/maps-platform
-
-    Once you have it, the GMAPS_API_KEY environmental variable will need to be set. In Windows this can be done
-    by running the command prompt in Administrator mode and running:
-
-    >> setx GMAPS_API_KEY "yourlonggooglemapsapikey"
-
-    If Anaconda or your Python environment is running you will need to restart it for the environmental variables to
-    take effect.
-
-    Additionally, the function requires the 'gmaps' library installed, instructions to install it are at the link below:
-    https://jupyter-gmaps.readthedocs.io/en/latest/install.html
-
-    :param meas_station:    A measurement station object which contains the latitude and longitude of a
-                            measurement location.
-    :type meas_station:     MeasurementStation
-    :param map_type:        Google maps base map types to use for the image. Google maps offers three different base
-                            map types: 'SATELLITE', 'HYBRID', 'TERRAIN'
-                            (see https://jupyter-gmaps.readthedocs.io/en/latest/tutorial.html)
-    :type map_type:         str
-    :param zoom_level:      Google maps zoom_level to use for the image.
-                            (see https://jupyter-gmaps.readthedocs.io/en/latest/tutorial.html)
-    :type zoom_level:       int
-    :return:                Google maps image with input measurement and reference nodes locations.
-    :rtype:                 fig
-
-    **Example usage**
-    ::
-        import brightwind as bw
-        mm1 = bw.MeasurementStation(bw.demo_datasets.demo_wra_data_model)
-        bw.plot_meas_loc_on_gmap(mm1)
-
-        bw.plot_meas_loc_on_gmap(mm1, map_type='SATELLITE')
-
-    """
-    gmaps.configure(api_key=utils.get_environment_variable('GMAPS_API_KEY'))
-    figure_layout = {
-        'width': '900px',
-        'height': '600px',
-        'margin': '0 auto 0 auto',
-        'padding': '1px'
-    }
-
-    # Plot meas locations
-    meas_loc_points = []
-    if isinstance(meas_station, MeasurementStation):
-        point = _create_coord_dict(meas_station.name, meas_station.lat, meas_station.long)
-        meas_loc_points.append(point)
-    else:
-        raise TypeError('Error with format of meas_loc, please input dataframe or dictionary or list of dictionaries')
-
-    # Assign center of figure as first meas_loc in list
-    fig = gmaps.figure(center=meas_loc_points[0]['coords'], map_type=map_type,
-                       zoom_level=zoom_level, layout=figure_layout)
-
-    for i_color, meas_loc_point in enumerate(meas_loc_points):
-        marker = gmaps.marker_layer([meas_loc_point['coords']],
-                                    info_box_content=meas_loc_point['name'],
-                                    display_info_box=True)
-        fig.add_layer(marker)
-    return fig
 
 
 def _replace_none_date(list_or_dict):
