+++ requirements.txt
@@ -9,3 +9,4 @@ six             >= 1.12.0
 python-dateutil >= 2.8.0
 ipywidgets      >= 7.4.2
 ipython         >= 7.4.0
+gmaps           >= 0.9.0
